created a common Requests class for all Controller tests and changed InvoiceControllerTest and  InvoiceControllerStepwiseTest
